V2.1.0
Had to do a lot of changes so that things don't overwrite each other. Did it so that when assists/leads are traded items it checks if the item was already added to the loot list. If it wasn't, it pops up asking. It uses a queue for the items so if you're traded more than one it still works (still doesn't work if you pick up more than one, but I don't think you can pick up more than one tradable items)