V2.7.3
--Bug fix for the list updating randomly.
I had a boolean comparison done wrong, so it wasn't getting an early out and was taking info from other addons that had "add" in the message. No more!

--Feature finally done! Addon now works when traded an item from an off-server player!
The addon originally looked for the text "You receive item: [item link]' and apparently that just doesn't even happen when you trade with someone off-server? It now looks for the event that puts out the text "Trade Completed" in the middle of your screen! It then uses that event to trigger going through your last traded items and for each item that meets the criteria it asks if you want to add it to the list.